make Input.swift file
